requirements.txt
pyyaml
numpy
pandas
openpyxl
python-dateutil
plotly
flask
flask_sqlalchemy
dash
python-dotenv